[refactor] OAuth and access token
