+++ http/oauth/google.go
@@ -14,27 +14,8 @@
  * limitations under the License.
  */
 
-package v1
+package oauth
 
-import (
-	"github.com/gin-gonic/gin"
-	"log"
-	"net/http"
-)
-
-func Error(msg string, c *gin.Context) {
-	c.JSON(http.StatusBadRequest, gin.H{"success": false, "message": msg})
-}
-
-func Success(data interface{}, c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{"success": true, "data": data})
-}
-
-func ErrorServer(err error, c *gin.Context) {
-	log.Println(err)
-	c.JSON(http.StatusInternalServerError, gin.H{"success": false, "message": "Internal Server Error"})
-}
-
-func ErrorUnauthorized(msg string, c *gin.Context) {
-	c.JSON(http.StatusUnauthorized, gin.H{"success": false, "message": msg})
+func HandleGoogleAuthToken(token string) (accessToken string, err error) {
+	return "", nil
 }
