+++ http/server/router.go
@@ -36,6 +36,7 @@ func NewRouter() *gin.Engine {
 		userGroup := v1g.Group("users")
 		userCtrl := new(v1.UserController)
 
+		userGroup.POST("/auth/google/:token", userCtrl.GoogleAuth)
 		userGroup.POST("/auth", userCtrl.Auth)
 		userGroup.POST("", userCtrl.Create)
